code/Backend_Server/notification_service/requirements.txt
fastapi
uvicorn
aio-pika
aiosmtplib
pydantic
python-dotenv
pydantic-settings
sqlalchemy
psycopg2-binary
python-jose[cryptography]
